requirements.txt
Django==4.2
django-heroku==0.3.1
djangorestframework==3.14.0
numpy==1.24.3
pandas==1.5.3
plotly==5.14.1
Requests==2.29.0
streamlit==1.21.0
scikit-learn==1.2.2